Move the logic to get the configurations for a list of projects from jest-cli to jest-config
