+++ packages/jest-config/src/index.js
@@ -15,6 +15,8 @@ import type {
   ProjectConfig,
 } from 'types/Config';
 
+import chalk from 'chalk';
+import fs from 'fs';
 import path from 'path';
 import {isJSONString, replaceRootDirInPath} from './utils';
 import normalize from './normalize';
@@ -85,7 +87,7 @@ export function readConfig(
   }
 
   const {options, hasDeprecationWarnings} = normalize(rawOptions, argv);
-  const {globalConfig, projectConfig} = getConfigs(options);
+  const {globalConfig, projectConfig} = groupOptions(options);
   return {
     configPath,
     globalConfig,
@@ -94,7 +96,7 @@ export function readConfig(
   };
 }
 
-const getConfigs = (
+const groupOptions = (
   options: Object,
 ): {globalConfig: GlobalConfig, projectConfig: ProjectConfig} => ({
   globalConfig: Object.freeze({
@@ -202,3 +204,115 @@ const getConfigs = (
     watchPathIgnorePatterns: options.watchPathIgnorePatterns,
   }),
 });
+
+const ensureNoDuplicateConfigs = (parsedConfigs, projects, rootConfigPath) => {
+  const configPathMap = new Map();
+
+  for (const config of parsedConfigs) {
+    const {configPath} = config;
+    if (configPathMap.has(configPath)) {
+      const message = `Whoops! Two projects resolved to the same config path: ${chalk.bold(
+        String(configPath),
+      )}:
+
+  Project 1: ${chalk.bold(projects[parsedConfigs.findIndex(x => x === config)])}
+  Project 2: ${chalk.bold(
+    projects[parsedConfigs.findIndex(x => x === configPathMap.get(configPath))],
+  )}
+
+This usually means that your ${chalk.bold(
+        '"projects"',
+      )} config includes a directory that doesn't have any configuration recognizable by Jest. Please fix it.
+`;
+
+      throw new Error(message);
+    }
+    if (configPath !== null) {
+      configPathMap.set(configPath, config);
+    }
+  }
+};
+
+// Possible scenarios:
+//  1. jest --config config.json
+//  2. jest --projects p1 p2
+//  3. jest --projects p1 p2 --config config.json
+//  4. jest --projects p1
+//  5. jest
+//
+// If no projects are specified, process.cwd() will be used as the default
+// (and only) project.
+export function readConfigs(
+  argv: Argv,
+  projectPaths: Array<Path>,
+): {
+  globalConfig: GlobalConfig,
+  configs: Array<ProjectConfig>,
+  hasDeprecationWarnings: boolean,
+} {
+  let globalConfig;
+  let hasDeprecationWarnings;
+  let configs: Array<ProjectConfig> = [];
+  let projects = projectPaths;
+  let configPath: ?Path;
+
+  if (projectPaths.length === 1) {
+    const parsedConfig = readConfig(argv, projects[0]);
+    configPath = parsedConfig.configPath;
+
+    if (parsedConfig.globalConfig.projects) {
+      // If this was a single project, and its config has `projects`
+      // settings, use that value instead.
+      projects = parsedConfig.globalConfig.projects;
+    }
+
+    hasDeprecationWarnings = parsedConfig.hasDeprecationWarnings;
+    globalConfig = parsedConfig.globalConfig;
+    configs = [parsedConfig.projectConfig];
+    if (globalConfig.projects && globalConfig.projects.length) {
+      // Even though we had one project in CLI args, there might be more
+      // projects defined in the config.
+      projects = globalConfig.projects;
+    }
+  }
+
+  if (projects.length > 1) {
+    const parsedConfigs = projects
+      .filter(root => {
+        // Ignore globbed files that cannot be `require`d.
+        if (
+          fs.existsSync(root) &&
+          !fs.lstatSync(root).isDirectory() &&
+          !root.endsWith('.js') &&
+          !root.endsWith('.json')
+        ) {
+          return false;
+        }
+
+        return true;
+      })
+      .map(root => readConfig(argv, root, true, configPath));
+
+    ensureNoDuplicateConfigs(parsedConfigs, projects, configPath);
+    configs = parsedConfigs.map(({projectConfig}) => projectConfig);
+    if (!hasDeprecationWarnings) {
+      hasDeprecationWarnings = parsedConfigs.some(
+        ({hasDeprecationWarnings}) => !!hasDeprecationWarnings,
+      );
+    }
+    // If no config was passed initially, use the one from the first project
+    if (!globalConfig) {
+      globalConfig = parsedConfigs[0].globalConfig;
+    }
+  }
+
+  if (!globalConfig || !configs.length) {
+    throw new Error('jest: No configuration found for any project.');
+  }
+
+  return {
+    configs,
+    globalConfig,
+    hasDeprecationWarnings: !!hasDeprecationWarnings,
+  };
+}
